Refactoring for promptgen tool
